+++ src/app/components/MyScreen.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function MyScreen() {
+  return (
+    <div className='bg-primary p-4 m-4 rounded-md'>
+      <div className="text-center">
+        <span className='font-bold text-2xl text-secondary'>My Screen</span>
+      </div>
+    </div>
+  )
+}
